% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_pipeline.R
\name{do_derep_step}
\alias{do_derep_step}
\title{Dereplication Step}
\usage{
do_derep_step(filt_Fs, filt_Rs)
}
\arguments{
\item{filt_Fs}{forward filtered reads}

\item{filt_Rs}{reverse filtered reads}
}
\value{
Returns list of errors

\describe{
  \item{\code{err_f}}{error rates learned for forward sequences}
  \item{\code{err_r}}{error rates learned for reverse sequences}
}
}
\description{
For DADA2 to work efficiently, samples sequences are dereplicated into
"unique sequences" with corresponding counts of these sequences.
}
\details{
This step reduces compute time by eliminating redundant comparisons.

**Note**: DADA retains quality information associated with each unique
sequence. It averages the sequences for each unique sequence. These quality
scores are used in the error model.

It is intended to take the output from `filterAndTrim()` into this function.
}
\examples{
example_map <- "mapping.txt"
example_data <- "sequences"
input_files <- do_input_step(example_map, example_data)
errors <- do_derep_step(input_files$seq_f, input_files$seq_r)
}
\references{
See "Dereplication"
  \url{http://benjjneb.github.io/dada2/tutorial.html#dereplication}
}
