% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_pipeline.R
\name{do_error_learn}
\alias{do_error_learn}
\title{Learn Error Rates for Sequences}
\usage{
do_error_learn(filt_Fs, filt_Rs)
}
\arguments{
\item{filt_Fs}{forward filtered reads}

\item{filt_Rs}{reverse filtered reads}
}
\value{
Returns list of errors

\describe{
  \item{\code{err_f}}{error rates learned for forward sequences}
  \item{\code{err_r}}{error rates learned for reverse sequences}
}
}
\description{
The error rates of the samples sequences are learned, so that they can be
accurately accounted for in the denoising step.
}
\examples{
example_map <- "mapping.txt"
example_data <- "sequences"
input_files <- do_input_step(example_map, example_data)
errors <- do_derep_step(input_files$seq_f, input_files$seq_r)
}
