% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_pipeline.R
\name{do_input_step}
\alias{do_input_step}
\title{Read in Data Step}
\usage{
do_input_step(map_file, data_path)
}
\arguments{
\item{map_file}{string with path to mapping file}

\item{data_path}{string with directory of samples}
}
\value{
returns list with four elements:

\describe{
  \item{\code{map}}{the mapping file}
  \item{\code{seq_f}}{forward sequencing sample paths}
  \item{\code{seq_r}}{reverse sequencing sample paths}
  \item{\code{seq_names}}{samples names for sequence samples}
}
}
\description{
Takes in strings for paths and returns a list to further manipulate in
downstream steps.
}
\examples{
example_map <- "mapping.txt"
example_data <- "sequences"
input_files <- do_input_step(example_map, example_data)
}
