% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_pipeline.R
\name{do_quality_step}
\alias{do_quality_step}
\title{Plot Quality Profile Step}
\usage{
do_quality_step(for_seqs, rev_seqs, plot_dir = "results", w = 7, h = 5)
}
\arguments{
\item{for_seqs}{character vector of paths for the forward sequencing samples}

\item{rev_seqs}{character vector of paths for the reverse sequencing samples}

\item{plot_dir}{string for directory name to output plots to}

\item{w}{width of the output figure}

\item{h}{height of the output figure}
}
\description{
Plot Quality Profile Step
}
\examples{
sample_f <- c("s001_R001.fastq", "s002_R001.fastq")
sample_r <- c("s001_R002.fastq", "s002_R002.fastq")
do_quality_step(sample_f, sample_r)
}
