% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_pipeline.R
\name{do_sample_infer_step}
\alias{do_sample_infer_step}
\title{Dereplicated Data and Infer Sample Sequence Variants}
\usage{
do_sample_infer_step(filt_Fs, err_f, filt_Rs, err_r, name_samples)
}
\arguments{
\item{filt_Fs}{filtered forward sequences}

\item{err_f}{error rates from filtered forward sequences}

\item{filt_Rs}{filtered forward sequences}

\item{err_r}{error rates from filtered reverse sequences}

\item{name_samples}{vector of sample names}
}
\value{
Returns list of

\describe{
  \item{\code{mergers}}{merged paired reads}
  \item{\code{dadaFs}}{inferred sequence variants for forward sequences}
  \item{\code{dadaRs}}{inferred sequence variants for reverse sequences}
}
}
\description{
Here is the main step in the DADA2 pipeline, which uses the `dada()`
function and dereplicate data.
}
\details{
For DADA2 to work efficiently, samples sequences are dereplicated into
"unique sequences" with corresponding counts of these sequences.

This step reduces compute time by eliminating redundant comparisons.

**Note**: DADA retains quality information associated with each unique
sequence. It averages the sequences for each unique sequence. These quality
scores are used in the error model.

It is intended to take the output filtered results from `filterAndTrim()`
into this function.
}
\examples{
do_sample_infer_step(filt_Fs, err_f, filt_Rs, err_r, name_samples)
}
\references{
See "Dereplication"
  \url{http://benjjneb.github.io/dada2/tutorial.html#dereplication}

See "Sample Inference"
  \url{http://benjjneb.github.io/dada2/tutorial.html#sample-inference}
}
