% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_pipeline.R
\name{do_taxonomy_step}
\alias{do_taxonomy_step}
\title{Construct Sequence Table, Remove Chimeras, and Assign Taxonomy}
\usage{
do_taxonomy_step(mergers, chim_method = "consensus", ref_db)
}
\arguments{
\item{mergers}{merged pairs of sequences}

\item{chim_method}{string for method of chimera removal, either "consensus"
(default) or "pooled"}
}
\value{
Returns list of

\describe{
  \item{\code{seqtab}}{sequence table}
  \item{\code{seqtab_nochim}}{sequence table without chimeras}
  \item{\code{taxa}}{assigned taxonomy from sequence variants}
}
}
\description{
Construct Sequence Table, Remove Chimeras, and Assign Taxonomy
}
\examples{
refdb <- "greengenes.fa.gz"
do_taxonomy_step(mergers, ref_db = refdb)
}
\references{
See "Construct sequence table"
  \url{http://benjjneb.github.io/dada2/tutorial.html#construct-sequence-table}

See "Remove chimeras"
  \url{http://benjjneb.github.io/dada2/tutorial.html#remove-chimeras}

See "Assign taxonomy"
  \url{http://benjjneb.github.io/dada2/tutorial.html#assign-taxonomy}
}
