% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_pipeline.R
\name{do_track_dada2}
\alias{do_track_dada2}
\title{Construct Sequence Table, Remove Chimeras, and Assign Taxonomy}
\usage{
do_track_dada2(out, dadaFs, mergers, seqtab, seqtab_nochim, sample_names)
}
\arguments{
\item{out}{matrix output from `filterAndTrim()` function}

\item{dadaFs}{dada object output from `dada()` function for forward reads}

\item{mergers}{output from `mergePairs()` function}

\item{seqtab}{sequence table output from `makeSequenceTable()`}

\item{seqtab_nochim}{output from `removeBimeraDenovo()`}

\item{sample_names}{vector of sample names}
}
\value{
matrix of sample rows and metric columns over the previous pipeline
  steps
}
\description{
Construct Sequence Table, Remove Chimeras, and Assign Taxonomy
}
\examples{
do_track_dada2(out, dadaFs, getN, mergers, seqtab, seqtab_nochim,
sample_names)
}
\references{
See "Track Reads Through the Pipeline"
  \url{http://benjjneb.github.io/dada2/tutorial.html#track-reads-through-the-pipeline}
}
