% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{extract_sample_id}
\alias{extract_sample_id}
\title{Extract Sample IDs}
\usage{
extract_sample_id(sample_list)
}
\arguments{
\item{sample_list}{vector of sample names}
}
\value{
vector of sample IDs
}
\description{
In a vector of sequence labels, return a vector of just the sample IDs with
the format of "s" followed by three digits.
}
\examples{
example_samples <- c("lane1-s216-index-GTTTCTGCT-s216.fastq.gz",
                     "lane1-s217-index-CACGATGGA-s217.fastq.gz")
extract_sample_id(example_samples)
}
