% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{filter_by_id}
\alias{filter_by_id}
\title{Filter Samples by ID}
\usage{
filter_by_id(ids, sample_list)
}
\arguments{
\item{ids}{vector of ID numbers or labels}

\item{sample_list}{vector of string sample names}
}
\value{
vector of original sample vector but filtered for certain IDs
}
\description{
Given a vector of sample names or labeling that has a sample ID, this
function will filter the sample names by IDs.
}
\examples{
example_samples <- c("lane1-s216-index-GTTTCTGCT-s216_L001_R1_001.fastq.gz",
                     "lane1-s217-index-CACGATGGA-s217_L001_R1_001.fastq.gz")
example_ids <- c("s216")
filter_by_id(example_ids, example_samples)
}
