% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{getN}
\alias{getN}
\title{Get Number of Unique Sequences}
\usage{
getN(x)
}
\arguments{
\item{x}{uniques vector from dada() output}
}
\value{
integer sum of number of unique sequences
}
\description{
Get Number of Unique Sequences
}
\examples{
dada_out <- dada(derep, err)
getN(dada_out[[1]])
}
