% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2_pipeline.R
\name{main}
\alias{main}
\title{DADA2 Automated Workflow}
\usage{
main()
}
\value{
NA
}
\description{
This workflow steams from the big data workflow tutorial listed in the DADA2
documentation.
}
\details{
This workflow should be run automatically from the root directory of this
repository using the Make utlity.
}
\examples{
main()
}
