% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_and_qc.R
\name{read_in_data}
\alias{read_in_data}
\title{Read in Sequence Data}
\usage{
read_in_data(path)
}
\arguments{
\item{path}{string with path to samples}
}
\value{
list with three elements: forward, reverse, and samples
}
\description{
This function reads in appropriate sequence data from a give directory. It
returns a list object that has the paths to all the samples and sample
names.
}
\examples{
path_to_samples <- "~/Samples/"
sample_paths <- read_in_data(path_to_samples)
}
