% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada2HPCPipe.R
\docType{package}
\name{dada2HPCPipe-package}
\alias{dada2HPCPipe}
\alias{dada2HPCPipe-package}
\title{dada2HPCPipe: DADA2 HPC Workflow Wrapper}
\description{
16S rRNA microbiome data analysis workflow using DADA2 and R on a high
performance cluster using Slurm. The functions here are wrappers around the
DADA2 functions in order to more efficiently use on a cluster. The functions
break down the workflow into parts based on stopping points that require a
user's input for the next steps.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/erictleung/dada2HPCPipe}
  \item Report bugs at \url{https://github.com/erictleung/dada2HPCPipe/issues}
}

}
\author{
\strong{Maintainer}: Eric Leung \email{leunge@ohsu.edu}

}
