% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_and_qc.R
\name{filter_and_trim}
\alias{filter_and_trim}
\title{Filter and Trim Sample Sequences}
\usage{
filter_and_trim(samplNames, fnFs, fnRs, outDir, truncLen)
}
\arguments{
\item{samplNames}{vector of sample names}

\item{fnFs}{vector of paths to forward samples}

\item{fnRs}{vector of paths to reverse samples}

\item{outDir}{specified directory to output filtered reads}

\item{truncLen}{a vector with two elements, the first being the truncated
length for the forward reads and the second being the truncated length for
the reverse reads}
}
\value{
new filtered reads put into the filtered/ directory
}
\description{
Filteres and trims paired forward and reverse *.fastq files
}
